ref(types): extract OrganizationStatsProps into a leaf module
types/overrides (a foundational override-registry type module) imported
`OrganizationStatsProps` from the heavy organizationStats/index view solely
for a type annotation, tethering that view + its subtree into the frontend
type-import SCC. Move the prop type to the dependency-light
organizationStats/types leaf and repoint importers (overrides + gsApp).

Largest type-import SCC: 1657 -> 1632.
