+++ static/app/views/organizationStats/index.tsx
@@ -1,6 +1,6 @@
 import {Component} from 'react';
 import styled from '@emotion/styled';
-import type {Location, LocationDescriptorObject} from 'history';
+import type {LocationDescriptorObject} from 'history';
 import omit from 'lodash/omit';
 import pick from 'lodash/pick';
 import moment from 'moment-timezone';
@@ -24,12 +24,11 @@ import {SentryDocumentTitle} from 'sentry/components/sentryDocumentTitle';
 import {DATA_CATEGORY_INFO, DEFAULT_STATS_PERIOD} from 'sentry/constants';
 import {t} from 'sentry/locale';
 import {ConfigStore} from 'sentry/stores/configStore';
-import {DataCategory, type DataCategoryInfo, type PageFilters} from 'sentry/types/core';
-import type {Organization} from 'sentry/types/organization';
+import {DataCategory, type DataCategoryInfo} from 'sentry/types/core';
 import type {Project} from 'sentry/types/project';
 import {decodeScalar} from 'sentry/utils/queryString';
 import {useLocation} from 'sentry/utils/useLocation';
-import {useNavigate, type ReactRouter3Navigate} from 'sentry/utils/useNavigate';
+import {useNavigate} from 'sentry/utils/useNavigate';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {
   canUseMetricsStatsBytesUI,
@@ -40,6 +39,7 @@ import {getPerformanceBaseUrl} from 'sentry/views/performance/utils';
 import {makeProjectsPathname} from 'sentry/views/projects/pathname';
 import {SettingsPageHeader} from 'sentry/views/settings/components/settingsPageHeader';
 
+import type {OrganizationStatsProps} from './types';
 import type {ChartDataTransform} from './usageChart';
 import {CHART_OPTIONS_DATACATEGORY} from './usageChart';
 import {UsageStatsOrganization as UsageStatsOrg} from './usageStatsOrg';
@@ -70,13 +70,6 @@ export const PAGE_QUERY_PARAMS = [
   // From show data discarded on client toggle
   'clientDiscard',
 ];
-
-export type OrganizationStatsProps = {
-  location: Location;
-  navigate: ReactRouter3Navigate;
-  organization: Organization;
-  selection: PageFilters;
-};
 
 export class OrganizationStatsInner extends Component<OrganizationStatsProps> {
   get dataCategoryInfo(): DataCategoryInfo {
