+++ static/app/views/organizationStats/types.tsx
@@ -1,4 +1,15 @@
-import type {SeriesApi} from 'sentry/types/organization';
+import type {Location} from 'history';
+
+import type {PageFilters} from 'sentry/types/core';
+import type {Organization, SeriesApi} from 'sentry/types/organization';
+import type {ReactRouter3Navigate} from 'sentry/utils/useNavigate';
+
+export type OrganizationStatsProps = {
+  location: Location;
+  navigate: ReactRouter3Navigate;
+  organization: Organization;
+  selection: PageFilters;
+};
 
 /**
  * Raw response from API endpoint
